refactor: security 리팩토링 (#121)
