--- server/src/main/java/seb/project/Codetech/global/auth/handler/UserAccessDeniedHandler.java
@@ -1,24 +1,0 @@
-package seb.project.Codetech.global.auth.handler;
-
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpStatus;
-import org.springframework.security.access.AccessDeniedException;
-import org.springframework.security.web.access.AccessDeniedHandler;
-import org.springframework.stereotype.Component;
-import seb.project.Codetech.global.auth.utils.ErrorResponder;
-
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-
-@Slf4j
-@Component
-public class UserAccessDeniedHandler implements AccessDeniedHandler {
-    @Override
-    public void handle(HttpServletRequest request, HttpServletResponse response, AccessDeniedException
-                       accessDeniedException) throws IOException, ServletException{
-        ErrorResponder.sendErrorResponse(response, HttpStatus.FORBIDDEN);
-        log.warn("Forbidden error happened: {}", accessDeniedException.getMessage());
-    }
-}
