--- server/src/main/java/seb/project/Codetech/global/auth/utils/ErrorResponder.java
@@ -1,19 +1,0 @@
-package seb.project.Codetech.global.auth.utils;
-
-import com.google.gson.Gson;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
-import seb.project.Codetech.global.response.ErrorResponse;
-
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-
-public class ErrorResponder {
-    public static void sendErrorResponse(HttpServletResponse response, HttpStatus status) throws IOException{
-        Gson gson = new Gson();
-        ErrorResponse errorResponse = ErrorResponse.of(status);
-        response.setContentType(MediaType.APPLICATION_JSON_VALUE);
-        response.setStatus(status.value());
-        response.getWriter().write(gson.toJson(errorResponse, ErrorResponse.class));
-    }
-}
