+++ server/src/main/java/seb/project/Codetech/security/auth/jwt/JwtTokenizer.java
@@ -1,0 +1,91 @@
+package seb.project.Codetech.security.auth.jwt;
+
+import java.nio.charset.StandardCharsets;
+import java.security.Key;
+import java.util.Calendar;
+import java.util.Date;
+import java.util.Map;
+
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Component;
+
+import io.jsonwebtoken.Claims;
+import io.jsonwebtoken.Jws;
+import io.jsonwebtoken.Jwts;
+import io.jsonwebtoken.io.Decoders;
+import io.jsonwebtoken.io.Encoders;
+import io.jsonwebtoken.security.Keys;
+import lombok.Getter;
+
+@Component
+public class JwtTokenizer {
+	@Getter
+	@Value("${jwt.secret-key}")
+	private String secretKey;
+
+	@Getter
+	@Value("${jwt.access-token-expiration-minutes}")
+	private int accessTokenExpirationMinutes;
+
+	@Getter
+	@Value("${jwt.refresh-token-expiration-minutes}")
+	private int refreshTokenExpirationMinutes;
+
+	public String encodeBase64SecretKey(String secretKey) {
+		return Encoders.BASE64.encode(secretKey.getBytes(StandardCharsets.UTF_8));
+	}
+
+	public String generateAccessToken(Map<String, Object> claims, String subject, Date expiration,
+		String base64EncodedSecretKey) {
+		Key key = getKeyFromBase64EncodedKey(base64EncodedSecretKey);
+
+		return Jwts.builder()
+			.setClaims(claims)
+			.setSubject(subject)
+			.setIssuedAt(Calendar.getInstance().getTime())
+			.setExpiration(expiration)
+			.signWith(key)
+			.compact();
+	}
+
+	private Key getKeyFromBase64EncodedKey(String base64EncodedSecretKey) {
+		byte[] keyBytes = Decoders.BASE64.decode(base64EncodedSecretKey);
+
+		return Keys.hmacShaKeyFor(keyBytes);
+	}
+
+	public String generateRefreshToken(String subject, Date expiration, String base64EncodedSecretKey) {
+		Key key = getKeyFromBase64EncodedKey(base64EncodedSecretKey);
+
+		return Jwts.builder()
+			.setSubject(subject)
+			.setIssuedAt(Calendar.getInstance().getTime())
+			.setExpiration(expiration)
+			.signWith(key)
+			.compact();
+	}
+
+	public Jws<Claims> getClaims(String jws, String base64EncodedSecretKey) {
+		Key key = getKeyFromBase64EncodedKey(base64EncodedSecretKey);
+
+		return Jwts.parserBuilder()
+			.setSigningKey(key)
+			.build().parseClaimsJws(jws);
+	}
+
+	public void verifySignature(String jws, String base64EncodedSecretKey) {
+		Key key = getKeyFromBase64EncodedKey(base64EncodedSecretKey);
+
+		Jwts.parserBuilder()
+			.setSigningKey(key)
+			.build()
+			.parseClaimsJws(jws);
+	}
+
+	public Date getTokenExpiration(int expirationMinutes) {
+		Calendar calendar = Calendar.getInstance();
+		calendar.add(Calendar.MINUTE, expirationMinutes);
+
+		return calendar.getTime();
+	}
+}
