+++ server/src/main/java/seb/project/Codetech/security/auth/utils/ErrorResponder.java
@@ -1,0 +1,22 @@
+package seb.project.Codetech.security.auth.utils;
+
+import java.io.IOException;
+
+import javax.servlet.http.HttpServletResponse;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
+
+import com.google.gson.Gson;
+
+import seb.project.Codetech.global.response.ErrorResponse;
+
+public class ErrorResponder {
+	public static void sendErrorResponse(HttpServletResponse response, HttpStatus status) throws IOException {
+		Gson gson = new Gson();
+		ErrorResponse errorResponse = ErrorResponse.of(status);
+		response.setContentType(MediaType.APPLICATION_JSON_VALUE);
+		response.setStatus(status.value());
+		response.getWriter().write(gson.toJson(errorResponse, ErrorResponse.class));
+	}
+}
