+++ server/src/main/java/seb/project/Codetech/security/oauth2/entity/Provider.java
@@ -1,0 +1,7 @@
+package seb.project.Codetech.security.oauth2.entity;
+
+public enum Provider {
+	GOOGLE,
+	KAKAO,
+	LOCAL;
+}
